+++ src/Yaat.Sim/Commands/CanonicalCommandType.cs
@@ -129,6 +129,8 @@ public enum CanonicalCommandType
     TemporaryAltitude,
     Cruise,
     OnHandoff,
+    Contact,
+    FrequencyChangeApproved,
 
     // Coordination
     CoordinationRelease,